Converts tests inside packages

+++ main_test.go
@@ -19,16 +19,26 @@ func init() {
 			cwd, err := os.Getwd()
 			Expect(err).NotTo(HaveOccurred())
 
-			pathToExecutable := filepath.Join(cwd, "ginkgo-convert")
-
-			cmd := exec.Command(pathToExecutable, "github.com/tjarratt/ginkgo-convert/fixtures")
-			err = cmd.Run()
-			Expect(err).NotTo(HaveOccurred())
-
-			convertedFile, err := ioutil.ReadFile(filepath.Join(cwd, "fixtures/xunit_ginkgo_test.go"))
+			rewriteTestsInPackage("github.com/tjarratt/ginkgo-convert/fixtures")
+			pathToFile := filepath.Join(cwd, "fixtures/xunit_ginkgo_test.go")
+			convertedFile, err := ioutil.ReadFile(pathToFile)
 			Expect(err).NotTo(HaveOccurred())
 
 			goldMaster, err := ioutil.ReadFile(filepath.Join(cwd, "fixtures/ginkgo_test_goldmaster.go"))
+			Expect(err).NotTo(HaveOccurred())
+			Expect(string(convertedFile)).To(Equal(string(goldMaster)))
+		})
+
+		It("rewrites all tests in your package", func() {
+			cwd, err := os.Getwd()
+			Expect(err).NotTo(HaveOccurred())
+
+			rewriteTestsInPackage("github.com/tjarratt/ginkgo-convert/fixtures")
+			pathToFile := filepath.Join(cwd, "fixtures/nested/nested_ginkgo_test.go")
+			convertedFile, err := ioutil.ReadFile(pathToFile)
+			Expect(err).NotTo(HaveOccurred())
+
+			goldMaster, err := ioutil.ReadFile(filepath.Join(cwd, "fixtures/nested/nested_goldmaster.go"))
 			Expect(err).NotTo(HaveOccurred())
 			Expect(string(convertedFile)).To(Equal(string(goldMaster)))
 		})
@@ -40,17 +50,37 @@ func killAllConvertedGinkgoTests() {
 	Expect(err).NotTo(HaveOccurred())
 
 	fixtures := filepath.Join(cwd, "fixtures")
-	dirContents, err := ioutil.ReadDir(fixtures)
+	killTestsInDir(fixtures)
+}
+
+func killTestsInDir(dir string) {
+	dirContents, err := ioutil.ReadDir(dir)
 	Expect(err).NotTo(HaveOccurred())
 
 	regex := regexp.MustCompile(".+_ginkgo_test.go")
 	for _, file := range dirContents {
+		if file.IsDir() {
+			killTestsInDir(filepath.Join(dir, file.Name()))
+			continue
+		}
+
 		if !regex.MatchString(file.Name()) {
 			continue
 		}
 
-		pathToFile := filepath.Join(fixtures, file.Name())
+		pathToFile := filepath.Join(dir, file.Name())
 		err = os.Remove(pathToFile)
 		Expect(err).NotTo(HaveOccurred())
 	}
+}
+
+func rewriteTestsInPackage(packageToRewrite string) (convertedFile []byte) {
+	cwd, err := os.Getwd()
+	Expect(err).NotTo(HaveOccurred())
+	pathToExecutable := filepath.Join(cwd, "ginkgo-convert")
+	cmd := exec.Command(pathToExecutable, packageToRewrite)
+	err = cmd.Run()
+	Expect(err).NotTo(HaveOccurred())
+
+	return
 }
